Move shim implementation to wrapper-utils and convert types to jsdoc
