+++ src/types/web-compat.ts
@@ -46,45 +46,3 @@ declare module "../features/web-compat.js" {
     request: import("@duckduckgo/messaging/lib/shared-types").MessagingBase<WebCompatMessages>['request']
   }
 }
-
-declare module "../content-feature.js" {
-  interface BaseStrictPropertyDescriptor {
-    configurable: boolean;
-    enumerable: boolean;
-  }
-
-  interface StrictDataDescriptor extends BaseStrictPropertyDescriptor {
-    value: any;
-    writable: boolean;
-  }
-
-  interface StrictGetDescriptor extends BaseStrictPropertyDescriptor {
-    get: () => any;
-  }
-
-  interface StrictSetDescriptor extends BaseStrictPropertyDescriptor {
-    set: (v: any) => void;
-  }
-
-  interface StrictGetSetDescriptor extends BaseStrictPropertyDescriptor {
-    get: () => any;
-    set: (v: any) => void;
-  }
-
-  export type StrictPropertyDescriptor = StrictDataDescriptor | StrictSetDescriptor | StrictGetDescriptor | StrictGetSetDescriptor;
-
-  type DefineInterfaceOptions = {
-    interfaceDescriptorOptions: StrictDataDescriptor;
-    constructorErrorMessage: string; // Error message to throw when constructor is called. Ignored when `disallowConstructor` is false
-    wrapToString: boolean; // Mask the `toString` representation of class methods, default is true
-  } & ({
-    allowConstructorCall: true; // Allow calling the constructor function without `new`, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-    disallowConstructor: false; // Disallow creating new instances, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-  } | {
-    allowConstructorCall: false; // Allow calling the constructor function without `new`, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-    disallowConstructor: true; // Disallow creating new instances, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-  } | {
-    allowConstructorCall: false; // Allow calling the constructor function without `new`, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-    disallowConstructor: false; // Disallow creating new instances, default is false. `allowConstructorCall` and `disallowConstructor` cannot be true at the same time
-  });
-}
